spurious output when running DTN2 under valgrind due to: 1) Threads shutdown uncleanly. 2) Lots of memory is intentionally freed by the OS on exit. environment variable "OASYS_CLEANUP_SINGLETONS" causes oasys to delete classes that inherit from Singleton on exit.
Patch provided by Amy Alford <aloomis@sarn.org>

+++ servlib/DTNServer.cc
@@ -280,8 +280,10 @@ DTNServer::close_datastore()
     ProphetStore::instance()->close();
     BundleStore::instance()->close();
     GlobalStore::instance()->close();
-
+    
+    if(!getenv("OASYS_CLEANUP_SINGLETONS")) {
     delete_z(store_);
+    }
 }
 
 void
@@ -303,6 +305,7 @@ DTNServer::shutdown()
     BundleDaemon::instance()->post_and_wait(new ShutdownRequest(), &done);
 
     DiscoveryTable::instance()->shutdown();
+    InterfaceTable::shutdown();
     close_datastore();
 }
